requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
python-dotenv
reflex==0.8.27
requests